Tests/LibWeb/Text/expected/HTML/media-can-play-type.txt

application/octet-stream: ""
video/webm: "probably"
audio/mpeg: "maybe"
invalid non-ascii: ""
